gemini/function_calling_bkp/requirements.txt
google-generativeai>=0.5.4